[Refactor] 출석체크 폴더구조 변경, 주석 추가 (25.02.03)
